python3/requirements.txt
avro-python3
confluent-kafka
kafka-python3